integrate kafka topic from ui components
