--- src/modules/Topics/components/TopicsTable/index.ts
@@ -1,1 +1,0 @@
-export * from './TopicsTable';
