+++ src/modules/Topics/components/index.ts
@@ -8,5 +8,4 @@ export * from './TopicAdvanceConfig';
 export * from './TopicDetailHead';
 export * from './TopicDetailView';
 export * from './UpdateTopicView';
-export * from './TopicsTable';
 export * from './CustomRetentionMessage';
